+++ src/hooks/index.js
@@ -1,0 +1,5 @@
+import UseListenResize from './UseListenResize';
+
+const a = null;
+
+export { UseListenResize, a };
